+++ CHANGELOG.md
@@ -6,9 +6,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 ### Added
+- feat: Add alerts providers (#114)
 ### Changed
+- chore(deps): Update dependencies
 ### Fixed
+- docs: Fix some readme typos
 ### Removed
+- fix: Remove mocks-server-plugin tag from package.json
 
 ## [2.0.5] - 2020-12-21